Move CNodeState, related structs, and State() fxn
from main.cpp to nodestate.h|.cpp.
Includes addition of doxygen comments and enforcing clang-format
